Berns 2.0: Performance improvements and refactoring
Version 2.0 is a mostly API-compatible refactor of all of the core methods that
make up Berns. The goal is to improve performance, mostly using mutable strings
and inlining variables that were otherwise short lived.

In addition, the target Ruby version has been raised to 2.5 or later. 2.4 has
reached its end of life.

See CHANGELOG.md for a benchmark and results.